Fix boost cast error when input missing; tidy up
