+++ CommandLine/emake/3FG.hpp
@@ -1,0 +1,8 @@
+#ifndef EMAKE_3FG_HPP
+#define EMAKE_3FG_HPP
+#include <string>
+#include "Game.hpp"
+
+bool read_3fg(const std::string &input_file, Game *game);
+
+#endif
