+++ CommandLine/emake/OptionsParser.cpp
@@ -90,7 +90,7 @@ OptionsParser::OptionsParser() : _desc("Options")
   _desc.add_options()
     ("help,h", "Print help messages")
     ("info,i", opt::value<std::string>(), "Provides a listing of Platforms, APIs and Extensions")
-    ("input",   opt::value<std::string>(), "Input game file")
+    ("input",   opt::value<std::string>()->default_value(""), "Input game file; currently, only test harness 3-file games (*.3fg) are supported. The --input string is optional.")
     ("output,o", opt::value<std::string>()->required(), "Output executable file")
     ("platform,p", opt::value<std::string>()->default_value(def_platform), "Target Platform (XLib, Win32, Cocoa)")
     ("workdir,w", opt::value<std::string>()->default_value(def_workdir), "Working Directory")
